% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarie.R
\name{ACP2}
\alias{ACP2}
\title{ACP}
\usage{
ACP2(X, y, d)
}
\arguments{
\item{X}{data.frame, with quantitative variable}

\item{y}{integer,  with class variable}

\item{d}{integer, number of axes wanted}
}
\value{
pca,  principal component analysis
}
\description{
ACP
}
\examples{
X=data.frame(fromage["calories"],fromage["magnesium"],fromage["lipides"],fromage["retinol"])
y=fromage$groupes.cah
d=3
ACP2(X,y,d)
}
