% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_pack.R
\name{acpGraph}
\alias{acpGraph}
\title{Graphique resultat kmeans couplé acp}
\usage{
acpGraph(vars, reel, resKM)
}
\arguments{
\item{vars}{DataFrame contenant les variables quantitatives}

\item{reel}{DataFrame contenant les variables quantitatives}

\item{resKM}{Resultats d'un clustering}
}
\value{

}
\description{
Graphique resultat kmeans couplé acp
}
