% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{barplotXY}
\alias{barplotXY}
\title{Barplot X versus Y}
\usage{
barplotXY(df, x, y)
}
\arguments{
\item{df}{dataframe which includes the categorical variable and the clusters variable}

\item{x}{name of the categorical variable}

\item{y}{name of the clusters variable}
}
\value{

}
\description{
Barplot X versus Y
}
