% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{barplotYX}
\alias{barplotYX}
\title{Diagramme à barres Y en fonction de X}
\usage{
barplotYX(df, x, y)
}
\arguments{
\item{df}{dataframe contenant la variable numerique et la variable classe d'appartenance}

\item{x}{nom de la variable catégorielle}

\item{y}{nom de la variable classe d'appartenance}
}
\value{

}
\description{
Diagramme à barres Y en fonction de X
}
