% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{boxplot}
\alias{boxplot}
\title{Boxplot}
\usage{
boxplot(df, x, y)
}
\arguments{
\item{df}{dataframe which includes the quantitative variable and the clusters variable}

\item{x}{name of the quantitative variable}

\item{y}{name of the clusters variable}
}
\value{

}
\description{
Boxplot
}
