% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesuresEval.R
\name{compareRes}
\alias{compareRes}
\title{Function which allows to compare the result of two clustering}
\usage{
compareRes(object1, object2)
}
\arguments{
\item{object1}{The first object following a first clustering}

\item{object2}{The second object following a second clustering}
}
\value{

}
\description{
Function which allows to compare the result of two clustering
}
\examples{
Obj2c <- EvalMetrics(pred_reel_2c$val_reel,pred_reel_2c$val_pred)
Obj2c_bis <- EvalMetrics(pred_reel_2c_bis$val_reel,pred_reel_2c_bis$val_pred)
compareRes(Obj2c,Obj2c_bis)
}
