% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_test.R
\docType{data}
\name{df_test}
\alias{df_test}
\title{Dataset df_test}
\format{
A data frame with 357 rows and 8 variables:
\describe{
\item{val_pred}{val_pred}
\item{val_reel}{val_reel}
...
}
}
\usage{
df_test
}
\description{
A dataset containing the test data
}
\keyword{datasets}
