% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{.vt.quali}
\alias{.vt.quali}
\title{Characterization of a clusters with a qualitative variable}
\usage{
.vt.quali(x, k, y)
}
\arguments{
\item{x}{qualitative variable}

\item{k}{name of the cluster selected}

\item{y}{clusters variable}
}
\value{

}
\description{
Characterization of a clusters with a qualitative variable
}
