% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{.vt.quanti}
\alias{.vt.quanti}
\title{Characterization of a cluster with a quantitative variable}
\usage{
.vt.quanti(x, k, y)
}
\arguments{
\item{x}{quantitative variable}

\item{k}{name of the cluster selected}

\item{y}{clusters variable}
}
\value{

}
\description{
Characterization of a cluster with a quantitative variable
}
