% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesuresEval.R
\name{EvalMetrics}
\alias{EvalMetrics}
\title{Création de l'objet "metrics" a partir de deux df contenant les valeurs prédites et réels}
\usage{
EvalMetrics(valreel, valpred)
}
\arguments{
\item{valreel}{Un dataframe contenant les classes réels}

\item{valpred}{Un dataframe contenant les classes prédites par un clustering}
}
\value{
Un objet de la classe metrics
}
\description{
Création de l'objet "metrics" a partir de deux df contenant les valeurs prédites et réels
}
