% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromage.R
\docType{data}
\name{fromage}
\alias{fromage}
\title{Dataset fromage}
\format{
A data frame with 29 rows and 10 variables:
\describe{
\item{calories}{presence of calories}
\item{sodium}{presence of sodium}
\item{calcium}{presence of calcium}
\item{lipides}{presence of lipides}
\item{retinol}{presence of retinol}
\item{folates}{presence of folates}
\item{proteines}{presence of proteines}
\item{cholesterol}{presence of cholesterol}
\item{magnesium}{presence of magnesium}
\item{groupes.cah}{class variable}
}
}
\usage{
fromage
}
\description{
A dataset containing the fromage data
}
\keyword{datasets}
