% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesuresEval.R
\name{ggMatConf}
\alias{ggMatConf}
\title{Confusion Matrix Graph}
\usage{
ggMatConf(objet)
}
\arguments{
\item{objet}{#Metrics variable}
}
\description{
Confusion Matrix Graph
}
\examples{
Obj2c <- EvalMetrics(pred_reel_2c$val_reel,pred_reel_2c$val_pred)
ggMatConf(Obj2c)
}
