% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarie.R
\name{LDA2}
\alias{LDA2}
\title{Linear Discriminant Analysis}
\usage{
LDA2(X, y)
}
\arguments{
\item{X}{data.frame, with quantitative variable}

\item{y}{integer,  with class variable}
}
\value{
lda, linear discriminant analysis
}
\description{
Linear Discriminant Analysis
}
\examples{
X=data.frame(fromage["calories"],fromage["magnesium"],fromage["lipides"],fromage["retinol"])
y=fromage$groupes.cah
LDA2(X,y)
}
