% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_reel_2c.R
\docType{data}
\name{pred_reel_2c}
\alias{pred_reel_2c}
\title{Dataset pred_reel_2c}
\format{
A data frame with 357 rows and 2 variables:
\describe{
\item {val_pred}{predicted  membership class variable}
\item {val_reel}{real membership class variable}
}
}
\usage{
pred_reel_2c
}
\description{
A dataset containing the value of prediction and the real values, 2 different values possible
}
\keyword{datasets}
