% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_reel_2c_bis.R
\docType{data}
\name{pred_reel_2c_bis}
\alias{pred_reel_2c_bis}
\title{Dataset pred_reel_2c_bis}
\format{
A data frame with 357 rows and 2 variables:
\describe{
\item val_pred,predicted  membership class variable
\item val_reel, real membership class variable
}
}
\usage{
pred_reel_2c_bis
}
\description{
A dataset containing the value of prediction and the real values, 2 different values possible
}
\keyword{datasets}
