% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_reel_3c.R
\docType{data}
\name{pred_reel_3c}
\alias{pred_reel_3c}
\title{Dataset pred_reel_3c}
\format{
A data frame with 357 rows and 2 variables:
\describe{
\item{val_pred}{val_pred}
\item{val_reel}{val_reel}
...
}
}
\usage{
pred_reel_3c
}
\description{
A dataset containing the value of prediction and the real values, 3 different values possible
}
\keyword{datasets}
