% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{print.cluster}
\alias{print.cluster}
\title{Surcharge de print pour l'affichage de l'objet "cluster"}
\usage{
\method{print}{cluster}(obj)
}
\arguments{
\item{obj}{instance de la classe cluster}
}
\value{

}
\description{
Surcharge de print pour l'affichage de l'objet "cluster"
}
