% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{radar}
\alias{radar}
\title{Radar diagram}
\usage{
radar(X, y)
}
\arguments{
\item{X}{one or more quantitative variables}

\item{y}{clusters variable}
}
\description{
Radar diagram
}
\examples{
df_test=as.data.frame(df_test)
data_quanti=df_test[,3:5]
radar(data_quanti,df_test$val_pred)
}
