% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{resCluster}
\alias{resCluster}
\title{Characterization of a selected cluster}
\usage{
resCluster(X, y, k)
}
\arguments{
\item{X}{all variables}

\item{y}{name of the cluster selected}

\item{k}{clusters variable}
}
\value{

}
\description{
Characterization of a selected cluster
}
