% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{tab.quali.ligne}
\alias{tab.quali.ligne}
\title{Table of row profiles for a categorical variable}
\usage{
tab.quali.ligne(x, y)
}
\arguments{
\item{x}{categorical variable}

\item{y}{clusters variable}
}
\value{
table
}
\description{
Table of row profiles for a categorical variable
}
\examples{
tab.quali.ligne(df_test$val_pred,df_test$sexe)
}
