% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{tab.quanti}
\alias{tab.quanti}
\title{Table of conditional means and \% explained variance}
\usage{
tab.quanti(X, y)
}
\arguments{
\item{X}{one or more quantitative variables}

\item{y}{clusters variable}
}
\value{

}
\description{
Table of conditional means and \% explained variance
}
