% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{vcramer}
\alias{vcramer}
\title{Cramer's V}
\usage{
vcramer(x, y)
}
\arguments{
\item{x}{categorical variable}

\item{y}{clusters variable}
}
\value{
a float
}
\description{
Cramer's V
}
\examples{
vcramer(df_test$sexe,df_test$val_pred)
}
