% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{vt.quali}
\alias{vt.quali}
\title{Caractérisation d'un groupe avec une variable qualitative}
\usage{
vt.quali(x, k, y)
}
\arguments{
\item{x}{variable qualitative}

\item{k}{cluster choisi}

\item{y}{variable classe d'appartenance}
}
\value{

}
\description{
Caractérisation d'un groupe avec une variable qualitative
}
