% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{vt.quanti}
\alias{vt.quanti}
\title{Caractérisation d'un groupe avec une variable quantitative}
\usage{
vt.quanti(x, k, y)
}
\arguments{
\item{x}{variable quantitative}

\item{k}{cluster choisi}

\item{y}{variable classe d'appartenance}
}
\value{

}
\description{
Caractérisation d'un groupe avec une variable quantitative
}
