% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarie.R
\name{acpGraph}
\alias{acpGraph}
\title{Graphique resultat kmeans couple acp}
\usage{
acpGraph(vars, reel, resKM)
}
\arguments{
\item{vars}{DataFrame with quantitatives variable}

\item{reel}{DataFrame with quantitatives variable}

\item{resKM}{clustering results}
}
\value{

}
\description{
Graphique resultat kmeans couple acp
}
