% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarie.R
\name{acpGraph}
\alias{acpGraph}
\title{Plot from ACP}
\usage{
acpGraph(X, y)
}
\arguments{
\item{X}{DataFrame with quantitatives variable}

\item{y}{clustering results}
}
\description{
Plot from ACP
}
\examples{
X=data.frame(fromage["calories"],fromage["magnesium"],fromage["lipides"],fromage["retinol"])
y=fromage$groupes.cah
acpGraph(X,y)
}
