% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{barplotYX}
\alias{barplotYX}
\title{Barplot Y versus X}
\usage{
barplotYX(df, x, y)
}
\arguments{
\item{df}{dataframe which includes the categorical variable and the clusters variable}

\item{x}{name of the categorical variable}

\item{y}{name of the clusters variable}
}
\description{
Barplot Y versus X
}
\examples{
barplotYX(df_test, "sexe", "val_pred")
}
