% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{boxplot}
\alias{boxplot}
\title{Boxplot}
\usage{
boxplot(df, x, y)
}
\arguments{
\item{df}{variables explicatives + classes}

\item{x}{nom var quanti}

\item{y}{nom var classe}
}
\value{

}
\description{
Boxplot
}
