% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_pack.R
\name{compareRes}
\alias{compareRes}
\title{Fonction qui permet de comparer le resultat de deux clustering}
\usage{
compareRes(object1, object2)
}
\arguments{
\item{object2}{}
}
\value{

}
\description{
Fonction qui permet de comparer le resultat de deux clustering
}
