% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_test.R
\docType{data}
\name{df_test}
\alias{df_test}
\title{Dataset df_test}
\format{
A data frame with 357 rows and 8 variables:
\describe{
\item val_pred, predicted value
\item val_reel, real values
\item taille, the height of the person
\item poids, the person's weight in kg
\item pointure, The shoe size
\item sexe, Man or woman
\item sport,sport practiced by the individual
\item val_parent, parent target value
}
}
\usage{
df_test
}
\description{
A dataset containing the test data
}
\keyword{datasets}
