% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesuresEval.R
\name{EvalMetrics}
\alias{EvalMetrics}
\title{Creation of the "metrics" object from two df containing the predicted and actual values}
\usage{
EvalMetrics(valreel, valpred)
}
\arguments{
\item{valreel}{A dataframe containing the real classes}

\item{valpred}{A dataframe containing the classes predicted by a clustering}
}
\value{
An object of the metrics class
}
\description{
Creation of the "metrics" object from two df containing the predicted and actual values
}
\examples{
EvalMetrics(pred_reel_2c$val_reel,pred_reel_2c$val_pred)

}
