% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_pack.R
\name{EvalMetrics}
\alias{EvalMetrics}
\title{Création de l'objet "metrics" a partir de deux df contenant les valeurs prédites et réels}
\usage{
EvalMetrics(valreel, valpred)
}
\arguments{
\item{valpred}{}
}
\value{

}
\description{
Création de l'objet "metrics" a partir de deux df contenant les valeurs prédites et réels
}
