% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{print.cluster}
\alias{print.cluster}
\title{Overload print function for displaying the "cluster" object}
\usage{
\method{print}{cluster}(obj)
}
\arguments{
\item{obj}{cluster class instance}
}
\value{

}
\description{
Overload print function for displaying the "cluster" object
}
