% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_pack.R
\name{print.metrics}
\alias{print.metrics}
\title{Surcharge de print pour l'affichage de l'objet "metrics"}
\usage{
\method{print}{metrics}(obj)
}
\arguments{
\item{obj}{}
}
\value{

}
\description{
Surcharge de print pour l'affichage de l'objet "metrics"
}
