% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{resCluster}
\alias{resCluster}
\title{Characterization of a selected cluster}
\usage{
resCluster(X, y, k)
}
\arguments{
\item{X}{all variables}

\item{y}{name of the cluster selected}

\item{k}{clusters variable}
}
\value{
object with class S3 "cluster"
}
\description{
Characterization of a selected cluster
}
\examples{
data=as.data.frame(df_test[,-c(1,2)])
resCluster(data,df_test$val_pred,"grand")
}
