% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{resCluster}
\alias{resCluster}
\title{Caractérisation d'un groupe choisi}
\usage{
resCluster(X, y, k)
}
\arguments{
\item{X}{toutes les vars}

\item{y}{cluster choisi}

\item{k}{variables classe d'appartenance}
}
\value{

}
\description{
Caractérisation d'un groupe choisi
}
