% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{tab.quali.col}
\alias{tab.quali.col}
\title{Table of column profiles for a categorical variable}
\usage{
tab.quali.col(x, y)
}
\arguments{
\item{x}{categorical variable}

\item{y}{clusters variable}
}
\value{
table
}
\description{
Table of column profiles for a categorical variable
}
\examples{
tab.quali.col(df_test$val_pred,df_test$sexe)
}
