% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{tab.quali.col}
\alias{tab.quali.col}
\title{Tableau des profils colonnes ligne pour une variable qualitative}
\usage{
tab.quali.col(x, y)
}
\arguments{
\item{x}{variable qualitative}

\item{y}{variable classe d'appartenance}
}
\value{

}
\description{
Tableau des profils colonnes ligne pour une variable qualitative
}
