% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{tab.quali.ligne}
\alias{tab.quali.ligne}
\title{Tableau des profils ligne pour une variable qualitative}
\usage{
tab.quali.ligne(x, y)
}
\arguments{
\item{x}{variable qualitative}

\item{y}{variable classe d'appartenance}
}
\value{

}
\description{
Tableau des profils ligne pour une variable qualitative
}
