% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{tab.quali}
\alias{tab.quali}
\title{Tableau des profils pour une variable qualitative}
\usage{
tab.quali(y, x)
}
\arguments{
\item{y}{variable classe d'appartenance}

\item{x}{variable qualitative}
}
\value{

}
\description{
Tableau des profils pour une variable qualitative
}
