% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{tab.quanti}
\alias{tab.quanti}
\title{Table of conditional means and \% explained variance}
\usage{
tab.quanti(X, y)
}
\arguments{
\item{X}{one or more quantitative variables}

\item{y}{clusters variable}
}
\value{
table
}
\description{
Table of conditional means and \% explained variance
}
\examples{
df_test=as.data.frame(df_test)
data_quanti=df_test[,3:5]
tab.quanti(data_quanti,df_test$val_pred)
}
