% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarie.r
\name{tab.quanti}
\alias{tab.quanti}
\title{Tableau des moyennes conditionnelles et \% de variance expliquée}
\usage{
tab.quanti(X, y)
}
\arguments{
\item{X}{une ou plusieurs vars quantitatives}

\item{y}{variable classe d'appartenance}
}
\value{

}
\description{
Tableau des moyennes conditionnelles et \% de variance expliquée
}
